pyproject.toml
[project]
name = "rag-apps"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.117.1",
    "langchain>=0.3.27",
    "langchain-community>=0.3.30",
    "langchain-groq>=0.3.8",
    "langchain-openai>=0.3.33",
    "python-dotenv>=1.1.1",
    "qdrant-client>=1.15.1",
    "requests>=2.32.5",
    "streamlit>=1.50.0",
    "uvicorn>=0.37.0",
]
